+++ packages/block-library/src/template-part/edit/index.js
@@ -264,9 +264,7 @@ export default function TemplatePartEdit( {
 									} );
 								} }
 							>
-								{ window?.__experimentalContentOnlyPatternInsertion
-									? __( 'Edit section' )
-									: __( 'Edit' ) }
+								{ __( 'Edit section' ) }
 							</ToolbarButton>
 						</BlockControls>
 					) }
@@ -324,7 +322,7 @@ export default function TemplatePartEdit( {
 					} }
 				</BlockSettingsMenuControls>
 
-				<InspectorControls>
+				<InspectorControls group="settings">
 					<TemplatesList
 						area={ area }
 						clientId={ clientId }