+++ packages/sync/src/index.ts
@@ -16,6 +16,7 @@ export * as Y from 'yjs';
  */
 export { default as Delta } from './quill-delta/Delta';
 
+export { AwarenessState } from './awareness/awareness-state';
 export {
 	CRDT_DOC_META_PERSISTENCE_KEY,
 	CRDT_RECORD_MAP_KEY,
@@ -27,4 +28,9 @@ export {
 	WORDPRESS_META_KEY_FOR_CRDT_DOC_PERSISTENCE,
 } from './config';
 export { createSyncManager } from './manager';
+
+/**
+ * An enhanced state includes additional metadata about the user's connection.
+ */
+export type { EnhancedState } from './awareness/awareness-types';
 export type * from './types';